Make a separate class for handling captured frames
Doing frame logic within the main Tmux class is pretty clunky.
It unnecessarily couples tmux invocation logic with processing
of results.

On one hand TmuxFrame is a UserString, so
it's essentially interchangeable
wherever older str result was used. On other hand, being a
custom class, it introduces a lot of useful methods, especially
for structural analysis of frame.

Signed-off-by: Vitaly Ostrosablin <tmp6154@yandex.ru>