TmuxFrame: Support VT100 ACS boxes.
There's no guarantee that ncurses would render it's boxes with
UTF-8 line-drawing characters. If it chooses to use VT100 ACS
characters, then parser would not work. Unfortunately, if it
decides to do so - tmux capture-frame would return plain ASCII
letters instead of proper boxes.

This commit (sort of) adds support for ACS boxes, although it's
not particularly versatile - we just add letters that could be
line drawing characters in corner/line candidates and consider
them as a part of line.

This works for NetHack, although it may very well break if used
for something else, since it's easy to introduce such letter
sequences anywhere. Also, we do not check for line consistency,
e.g. if horizontal line gets q in middle - it'd be considered
a part of that line.

So, it's more a quick fix and in makes code brittler. But this
method is already more complex than I like, so let's keep it
that way unless I find a good reason not to.

Signed-off-by: Vitaly Ostrosablin <tmp6154@yandex.ru>